Add tests and docs placeholders
